requirement.txt
ultralytics
streamlit
pillow
numpy